- resolve #332: Complete cleanup of the Request DTO constructors that are based on specialized framework types
- resolve #332: Complete cleanup of the Request DTO constructors that are based on specialized framework types
