+++ Rollbar/DTOs/Request.cs
@@ -1,25 +1,6 @@
 namespace Rollbar.DTOs
 {
-    using Rollbar.Diagnostics;
     using System.Collections.Generic;
-    using System.Net.Http;
-    using System.Linq;
-    using Rollbar.Common;
-
-#if (NETSTANDARD || NETCOREAPP)
-    using Microsoft.AspNetCore.Http;
-#endif
-
-#if (NETCOREAPP)
-    using System.IO;
-    using System.Text;
-#endif
-
-#if (NETFX)
-    using System.ServiceModel.Channels;
-    using System.Web;
-    using Newtonsoft.Json;
-#endif
 
     /// <summary>
     /// Models Rollbar HTTP Request DTO.
@@ -52,120 +33,6 @@ namespace Rollbar.DTOs
             : base(arbitraryKeyValuePairs)
         {
 
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="Request"/> class.
-        /// </summary>
-        /// <param name="arbitraryKeyValuePairs">The arbitrary key value pairs.</param>
-        /// <param name="rollbarConfig">The rollbar configuration.</param>
-        public Request(
-            IDictionary<string, object> arbitraryKeyValuePairs
-            , IRollbarConfig rollbarConfig
-            )
-            : this(arbitraryKeyValuePairs, rollbarConfig, null)
-        {
-
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="Request" /> class.
-        /// </summary>
-        /// <param name="arbitraryKeyValuePairs">The arbitrary key value pairs.</param>
-        /// <param name="rollbarConfig">The rollbar configuration.</param>
-        /// <param name="httpRequest">The HTTP request.</param>
-        public Request(
-            IDictionary<string, object> arbitraryKeyValuePairs
-            , IRollbarConfig rollbarConfig
-            , HttpRequestMessage httpRequest
-            ) 
-            : base(arbitraryKeyValuePairs)
-        {
-            if (httpRequest != null)
-            {
-                Assumption.AssertNotNull(rollbarConfig, nameof(rollbarConfig));
-                this.SnapProperties(httpRequest, rollbarConfig);
-            }
-        }
-
-        private void SnapProperties(HttpRequestMessage httpRequest, IRollbarConfig rollbarConfig)
-        {
-            Assumption.AssertNotNull(httpRequest, nameof(httpRequest));
-
-            this.Url = httpRequest.RequestUri?.AbsoluteUri;
-            this.QueryString = httpRequest.RequestUri?.Query;
-            this.Params = null;
-
-            this.Headers = new Dictionary<string, string>(httpRequest.Headers.Count());
-            foreach (var header in httpRequest.Headers)
-            {
-                this.Headers.Add(header.Key, StringUtility.Combine(header.Value, ", "));
-            }
-
-            this.Method = httpRequest.Method.Method;
-            switch(this.Method.ToUpperInvariant())
-            {
-                case "POST":
-                    var task = httpRequest.Content.ReadAsStringAsync();
-                    task.Wait();
-                    this.PostBody = task.Result;
-                    this.PostParams = null;
-                    break;
-                case "GET":
-                    this.GetParams = null;
-                    break;
-                default:
-                    System.Diagnostics.Trace.WriteLine(
-                        $"No-op processing {this.Method.ToUpperInvariant()} HTTP method."
-                        );
-                    break;
-            }
-
-#if (NETFX)
-            string userIP = null;
-            const string HttpContextProperty = "MS_HttpContext";
-            const string RemoteEndpointMessagePropery = "System.ServiceModel.Channels.RemoteEndpointMessageProperty";
-            if (httpRequest.Properties.ContainsKey(HttpContextProperty))
-            {
-                HttpContextBase ctx = httpRequest.Properties[HttpContextProperty] as HttpContextBase;
-                if (ctx != null)
-                {
-                    userIP = ctx.Request.UserHostAddress;
-                }
-            }
-            else if (httpRequest.Properties.ContainsKey(RemoteEndpointMessagePropery))
-            {
-                RemoteEndpointMessageProperty remoteEndpoint = 
-                    httpRequest.Properties[RemoteEndpointMessagePropery] as RemoteEndpointMessageProperty;
-                if (remoteEndpoint != null)
-                {
-                    userIP = remoteEndpoint.Address;
-                }
-            }
-            this.UserIp = 
-                Request.DecideCollectableUserIPValue(userIP, rollbarConfig.IpAddressCollectionPolicy);
-#endif
-        }
-
-        private static string DecideCollectableUserIPValue(string initialIPAddress, IpAddressCollectionPolicy ipAddressCollectionPolicy)
-        {
-            switch(ipAddressCollectionPolicy)
-            {
-                case IpAddressCollectionPolicy.Collect:
-                    return initialIPAddress;
-                case IpAddressCollectionPolicy.CollectAnonymized:
-                    return IpAddressUtility.Anonymize(initialIPAddress);
-                case IpAddressCollectionPolicy.DoNotCollect:
-                    return null;
-                default:
-                    Assumption.FailValidation(
-                        "Unexpected enumeration value!"
-                        , nameof(ipAddressCollectionPolicy)
-                        );
-                    break;
-            }
-
-            return null;
         }
 
         /// <summary>
