+++ Rollbar/_RollbarPackagingStrategies.cd
@@ -18,7 +18,7 @@
   <Class Name="Rollbar.RollbarPackageDecoratorBase">
     <Position X="8.5" Y="4.5" Width="7.75" />
     <TypeIdentifier>
-      <HashCode>AAAABAAAAAAAAAAAAgAAAAAAAAAAAAAAgAAAQAAwAAA=</HashCode>
+      <HashCode>AAAABAAAAAAAAAAAAgAAAAAAAAAAAAAAgAAAQACwAAA=</HashCode>
       <FileName>RollbarPackageDecoratorBase.cs</FileName>
     </TypeIdentifier>
     <Lollipop Position="0.2" />
@@ -30,32 +30,11 @@
       <FileName>PersonPackageDecorator.cs</FileName>
     </TypeIdentifier>
   </Class>
-  <Class Name="Rollbar.AspNet.Mvc.ExceptionContextPackage">
-    <Position X="0.5" Y="11.25" Width="6" />
-    <TypeIdentifier>
-      <HashCode>AAAABAAAAAAAAAAAAAAAAAAAAAABAAAAEAAAAAAAAAA=</HashCode>
-      <FileName>AspNet\Mvc\ExceptionContextPackage.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="Rollbar.AspNet.Mvc.HttpRequestPackageDecorator">
-    <Position X="8.5" Y="10.75" Width="9" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAAAAAAAAAAACAAAAAAAAAAAAAAAgAAA=</HashCode>
-      <FileName>AspNet\Mvc\HttpRequestPackageDecorator.cs</FileName>
-    </TypeIdentifier>
-  </Class>
   <Class Name="Rollbar.MessagePackage">
     <Position X="0.5" Y="5" Width="6" />
     <TypeIdentifier>
       <HashCode>IAAAJAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAQAAAAAAA=</HashCode>
       <FileName>MessagePackage.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="Rollbar.AspNet.Mvc.HttpContextPackageDecorator">
-    <Position X="8.5" Y="13.25" Width="9" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAAAAAAAACAAAAAAAAAAAgAAAAAAgAAA=</HashCode>
-      <FileName>AspNet\Mvc\HttpContextPackageDecorator.cs</FileName>
     </TypeIdentifier>
   </Class>
   <Class Name="Rollbar.BodyPackage">
@@ -73,7 +52,7 @@
     </TypeIdentifier>
   </Class>
   <Class Name="Rollbar.CustomKeyValuePackageDecorator">
-    <Position X="8.5" Y="16" Width="9" />
+    <Position X="8.5" Y="10.5" Width="9" />
     <TypeIdentifier>
       <HashCode>AAAAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAgAAA=</HashCode>
       <FileName>CustomKeyValuePackageDecorator.cs</FileName>
@@ -84,6 +63,13 @@
     <TypeIdentifier>
       <HashCode>AAAAJAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAIAAAA=</HashCode>
       <FileName>ObjectPackage.cs</FileName>
+    </TypeIdentifier>
+  </Class>
+  <Class Name="Rollbar.HttpRequestMessagePackageDecorator">
+    <Position X="8.5" Y="13" Width="9" />
+    <TypeIdentifier>
+      <HashCode>AAAAAAAAAAAAAAAAAAAAAAAAAAABQAAAAAACAAAgABA=</HashCode>
+      <FileName>HttpRequestMessagePackageDecorator.cs</FileName>
     </TypeIdentifier>
   </Class>
   <Interface Name="Rollbar.IRollbarPackage">
